+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Add `zng::widget::VAR_EDITOR`.
 * Add `dyn AnyConfig::get_raw_serde_bidi`.
 * Add `AnyVarValue::eq_any`.
 * Add `zng::config::settings`.